# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 3.0.0
  description: Carter's Care v3 - NDIS & Aged Care Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /timesheets:
    get:
      operationId: getTimesheets
      tags: [timesheets]
      summary: List timesheets
      parameters:
        - name: staffId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: weekStart
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of timesheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Timesheet"
    post:
      operationId: createTimesheet
      tags: [timesheets]
      summary: Create timesheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimesheetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}:
    get:
      operationId: getTimesheet
      tags: [timesheets]
      summary: Get timesheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timesheet detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
    patch:
      operationId: updateTimesheet
      tags: [timesheets]
      summary: Update timesheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimesheetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}/approve:
    post:
      operationId: approveTimesheet
      tags: [timesheets]
      summary: Approve timesheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"

  # CASE NOTES
...
components:
  schemas:
...
    Timesheet:
      type: object
      required: [id, staffId, weekStart, weekEnd, totalHours, status, createdAt]
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        weekStart:
          type: string
        weekEnd:
          type: string
        totalHours:
          type: number
        regularHours:
          type: number
        overtimeHours:
          type: number
        allowances:
          type: number
        status:
          type: string
          enum: [draft, submitted, approved, rejected, paid]
        approvedById:
          type: ["integer", "null"]
        approvedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    TimesheetInput:
      type: object
      required: [staffId, weekStart, weekEnd]
      properties:
        staffId:
          type: integer
        weekStart:
          type: string
        weekEnd:
          type: string
        notes:
          type: ["string", "null"]
    TimesheetUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [draft, submitted, approved, rejected, paid]
        notes:
          type: ["string", "null"]
...